requirements.txt
aws-cdk-lib==2.149.0
boto3==1.34.146
constructs>=10.0.0,<11.0.0
